api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.2
pydantic==2.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
openai==1.30.1
